+++ src/events/AddLineItemCustomPropertiesEvent.php
@@ -3,7 +3,7 @@ namespace fostercommerce\klaviyoconnect\events;
 
 use yii\base\Event;
 
-class AddOrderLineItemDetailsEvent extends Event
+class AddLineItemCustomPropertiesEvent extends Event
 {
     public $properties = array();
     public $order = null;