+++ src/events/AddOrderCustomPropertiesEvent.php
@@ -3,9 +3,9 @@ namespace fostercommerce\klaviyoconnect\events;
 
 use yii\base\Event;
 
-class TrackEventMappingEvent extends Event
+class AddOrderCustomPropertiesEvent extends Event
 {
-    public $name;
-    public $eventId;
     public $properties = array();
+    public $order = null;
+    public $event = null;
 }