+++ src/events/AddProfileMappingEvent.php
@@ -1,9 +1,0 @@
-<?php
-namespace fostercommerce\klaviyoconnect\events;
-
-use yii\base\Event;
-
-class AddProfileMappingEvent extends Event
-{
-    public $mappings = array();
-}